+++ Runtime/Component/ServiceLocateComponent.cs
@@ -10,7 +10,7 @@ namespace SymphonyFrameWork.Utility
     /// </summary>
     [HelpURL("https://www.notion.so/SymphonyLocate-19d7c2c6cc02809ea815c3a750fa95ca?pvs=4")]
     [DefaultExecutionOrder(-1000)] // 最初に実行されるようにする。
-    public sealed class SymphonyLocate : MonoBehaviour
+    public sealed class ServiceLocateComponent : MonoBehaviour
     {
         [SerializeField]
         [Tooltip("ロケートするコンポーネント")]