+++ Runtime/Orchestrator/Internal/SymphonyLifetimeComponent.cs
@@ -5,7 +5,7 @@ namespace SymphonyFrameWork.Orchestrator
     /// <summary>
     ///     SymphonyOrchestratorのランタイム上のライフタイムを表すコンポーネントです。
     /// </summary>
-    internal sealed class SymphonyOrchestratorObject : MonoBehaviour
+    internal sealed class SymphonyLifetimeComponent : MonoBehaviour
     {
     }
 }